reorganizing and setters
