+++ vulcanai2/models/__init__.py
@@ -7,6 +7,7 @@ This file will allow for direct imports
 
 from .cnn import CNN
 from .dnn import DNN
+from .BaseNetwork import BaseNetwork
 
 
 __all__ = ['CNN'] #pretty sure this will be useful later
